Feature/Fix/Improve: Gameplay Enhancements and Random Item System (#44)
* Add random item in group

* Setup for random item

* Add item random item spawning

* Update RandomGroup.cs

* Add prefab for random item

* Add Dropbox

* Add Prefab

* Add trigger collect box, spawn item in collect box

* Add Shader outline to collect box

* Add sound when open box

TODO: Use something to notify player like HUD button to open box, not auto trigger collect like old ?
references: https://www.youtube.com/watch?v=5wBon8X_MaQ
Number 4

* Add sin movement

* Add Object Shake

* WIP

* WIP

* Install Dotween

* Add outline, Add healing player feature

* Add Custom VFX

* Fix world 1 conflict

* Setup Spawn Item

* Add draw gizmos button inspector

* Setup Drop Box spawning in world 1

* Update RandomGroupManager.cs

* Add Icon Setup

* Change read and write settings, add drop box sync collect effect

* Fix Collider, Fix Collider Collect not working

* Add Gun icon

* Fix cannot reload when ammo remain is not enough for reload

* Fix item cannot collect when drop

* Add ui polishing when gun data change

* Update config

+++ Assets/BattleField/Scripts/Core/Inventory/ItemConfig/HealthItemConfig.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Health", menuName = "Config/Health")]
 public class HealthItemConfig : ItemConfig<HealingItemType>
 {
-    public float healthAmount;
+    public byte healthAmount;
     public float usingTime;
 
     public override void ShowDebug()
